+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.1.1
+
+- fix: RxJS peerDependency warning when installing the package, closes[#205](https://github.com/MurhafSousli/ngx-highlightjs/issues/205) in [87c8521](https://github.com/MurhafSousli/ngx-highlightjs/pull/208/commits/87c85211d47b063b0e7cc1c4254f019708ca2f2a).
+- fix: Passing empty string to `[highlight]` input should clear the highlighted content (test cases is added for that), closes [#119](https://github.com/MurhafSousli/ngx-highlightjs/issues/119) in [2d67fb3](https://github.com/MurhafSousli/ngx-highlightjs/pull/208/commits/2d67fb34bdb268b27aab525a437b59eb2a839da2).
+
 ## 6.1.0
 
 - feat: Load highlighting theme dynamically, ability to switch between themes easily, closes [#194](https://github.com/MurhafSousli/ngx-highlightjs/issues/194) in [9422d29](https://github.com/MurhafSousli/ngx-highlightjs/commit/9422d295ab7254e5cf908c35e984efd2c801e8ec).